finally done it

+++ samples/cube-fn.py
@@ -1,4 +1,4 @@
 def cube(num) :
     return num*num*num
-num=int(input("type any no. "))
+num=float(input("type any no. "))
 print(cube(num))
